requirements.txt
PySide6>=6.6
Pillow>=10.0
anthropic>=0.40
